requirements.txt
wheel
pandas
psutil
python-dateutil
orjson
brotli
sqlparams
sqlalchemy

# # Optional dependencies
# boto3
# redshift_connector
# PyAthena>=2.1.0
# PyAthenaJDBC
